Extract a GrDynamicRectanize class from GrDynamicAtlas
Allows us to use the atlas layout code without the texture proxy stuff.
